feat: remove search from head bar to writers page, add 'Writers of Belarus' to head bar
