--- src/components/search/index.js
@@ -1,3 +1,0 @@
-import Search from "./search";
-
-export default Search;
