+++ src/components/writers/search/index.js
@@ -1,0 +1,3 @@
+import Search from "./search";
+
+export default Search;
