+++ src/components/writers/writers.js
@@ -6,7 +6,8 @@ import Divider from '@material-ui/core/Divider';
 import ListItemText from '@material-ui/core/ListItemText';
 import ListItemAvatar from '@material-ui/core/ListItemAvatar';
 import Avatar from '@material-ui/core/Avatar';
-import Typography from '@material-ui/core/Typography';
+
+import Search from './search'
 
 const useStyles = makeStyles(theme => ({
     header: {
@@ -26,23 +27,24 @@ const Writers = (props) => {
 
   return (
     <div className={classes.header}>
-    <List component="nav" aria-label="main mailbox folders">
-        {writersInfo.map((info) => (
-            <>
-            <Divider variant="inset" component="li" />
-            <ListItem button>
-                <ListItemAvatar>
-                    <Avatar alt={info.fullName} src={info.photo.url} />
-                </ListItemAvatar>
-                <ListItemText
-                primary={info.fullName}
-                secondary={info.city}
-                />
-            </ListItem>
-            <Divider variant="inset" component="li" />
-            </>
-        ))}
-    </List>
+        <Search />
+        <List component="nav" aria-label="main mailbox folders">
+            {writersInfo.map((info) => (
+                <>
+                <Divider variant="inset" component="li" />
+                <ListItem button>
+                    <ListItemAvatar>
+                        <Avatar alt={info.fullName} src={info.photo.url} />
+                    </ListItemAvatar>
+                    <ListItemText
+                    primary={info.fullName}
+                    secondary={info.city}
+                    />
+                </ListItem>
+                <Divider variant="inset" component="li" />
+                </>
+            ))}
+        </List>
   </div>
   );
 }
